requirements.txt
Flask
python-dotenv
uvicorn
ibm-watsonx-ai
asgiref
flask-cors